Launch Pad added + minor fixes to gravity gun + prefab spawner

+++ EiComponent/Utils/EiRandom.cs
@@ -41,9 +41,9 @@ namespace Eitrum
 			}
 		}
 
-		public float _Int {
+		public int _Int {
 			get {
-				return (float)random.Next ();
+				return random.Next ();
 			}
 		}
